made scatter test more general, for fasta and contigset
